+++ python_modules/airline-demo/airline_demo/types.py
@@ -1,0 +1,39 @@
+"""Type definitions for the airline_demo."""
+
+from collections import namedtuple
+
+from pyspark.sql import (
+    DataFrame,
+)
+from dagster import (
+    types,
+)
+
+AirlineDemoResources = namedtuple(
+    'AirlineDemoResources',
+    ('spark', 's3', 'db_url', 'db_engine', 'db_dialect', 'redshift_s3_temp_dir'),
+)
+
+SparkDataFrameType = types.PythonObjectType(
+    'SparkDataFrameType',
+    python_type=DataFrame,
+    description='A Pyspark data frame.',
+)
+
+# SqlAlchemyQueryType = types.PythonObjectType(
+#     'SqlAlchemyQueryType',
+#     python_type=sqlalchemy.orm.query.Query,
+#     description='A SQLAlchemy query.',
+# )
+
+# SqlAlchemySubqueryType = types.PythonObjectType(
+#     'SqlAlchemySubqueryType',
+#     python_type=sqlalchemy.sql.expression.Alias,
+#     description='A SQLAlchemy subquery',
+# )
+
+# SqlAlchemyResultProxyType = types.PythonObjectType(
+#     'SqlAlchemyResultProxyType',
+#     python_type=sqlalchemy.engine.ResultProxy,
+#     description='A SQLAlchemy result proxy',
+# )
